数据类重构
